requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
psycopg2==2.9.9
SQLAlchemy==2.0.31
opencv-python==4.10.0.84
Flask-Migrate==4.0.7
google-generativeai==0.7.2
boto3==1.34.143
python-dotenv==1.0.1
Flask-SocketIO==5.3.6
numpy